Move existing account methods to new concern
